runtime: split out input facilities into an `input` module.
