+++ src/generate/templates/imports.rs
@@ -1,6 +1,5 @@
-use gll::runtime::{
-    nd::Arrow, traverse, CodeLabel, ParseNode, ParseNodeKind, ParseNodeShape, Range,
-};
+use gll::parse_node::ParseNodeShape;
+use gll::runtime::{nd::Arrow, traverse, CodeLabel, ParseNode, ParseNodeKind};
 use std::any;
 use std::fmt;
 use std::marker::PhantomData;
